Safari-609.1.11.1/LayoutTests/http/tests/is-logged-in/unavailable-in-insecure-contexts.html

<!DOCTYPE html>
<html lang="en">
<head>
    <meta charset="UTF-8">
    <script src="/js-test-resources/js-test.js"></script>
</head>
<body>
<script>
    description("Tests that IsLoggedIn is unavailable in insecure contexts.");

    if (internals)
        internals.markContextAsInsecure();

    if (navigator.setLoggedIn === undefined)
        testPassed("navigator.setLoggedIn is undefined on a page with protocol " + document.location.protocol);
    else
        testFailed("navigator.setLoggedIn is defined and is of type " + typeof navigator.setLoggedIn + " on a page with protocol " + document.location.protocol);
</script>
</body>
</html>
