Safari-611.1.10/LayoutTests/http/tests/is-logged-in/unavailable-in-insecure-contexts-expected.txt

Tests that IsLoggedIn is unavailable in insecure contexts.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS navigator.setLoggedIn is undefined on a page with protocol http:
PASS successfullyParsed is true

TEST COMPLETE
